backend/requirements.txt
django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1


